requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.58.0
google-generativeai>=0.8.0
anthropic>=0.40.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
testcontainers[postgres]>=4.9.0
